client/requirements.txt
python-chess
kivy
berserk
pyserial==3.4
fentoboardimage